chatbot_requirements.txt
# Requirements for the chatbot feature
# Install with: pip install -r chatbot_requirements.txt

streamlit>=1.28.0
openai<1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

